second attempt submission

+++ src/main/java/com/example/demo/service/InhousePartService.java
@@ -14,7 +14,7 @@ import java.util.List;
  */
 public interface InhousePartService {
     public List<InhousePart> findAll();
-    public InhousePart findById(int theId);
+    public InhousePart findById(Long theId);
     public void save (InhousePart thePart);
-    public void deleteById(int theId);
+    public void deleteById(Long theId);
 }
